requirements.txt
chainlit
langchain
langchain-google-genai
langgraph
playwright
beautifulsoup4
python-dotenv
pytest
pytest-asyncio
langfuse
pydantic